feat: circuit loading and tests
