+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "libp2p",
-  "version": "0.12.4",
+  "version": "0.12.0",
   "description": "JavaScript base class for libp2p bundles",
   "main": "src/index.js",
   "scripts": {
@@ -39,11 +39,11 @@
   "dependencies": {
     "async": "^2.5.0",
     "libp2p-ping": "~0.6.0",
-    "libp2p-swarm": "~0.32.4",
-    "mafmt": "^3.0.1",
-    "multiaddr": "^3.0.1",
-    "peer-book": "~0.5.1",
-    "peer-id": "~0.10.1",
+    "libp2p-swarm": "~0.32.0",
+    "mafmt": "^3.0.0",
+    "multiaddr": "^3.0.0",
+    "peer-book": "~0.5.0",
+    "peer-id": "~0.10.0",
     "peer-info": "~0.11.0"
   },
   "devDependencies": {
@@ -51,39 +51,28 @@
     "chai": "^4.1.2",
     "dirty-chai": "^2.0.1",
     "cids": "~0.5.1",
-    "libp2p-kad-dht": "~0.5.1",
-    "libp2p-mdns": "~0.9.1",
+    "libp2p-kad-dht": "~0.5.0",
+    "libp2p-mdns": "~0.9.0",
     "libp2p-multiplex": "~0.5.0",
-    "libp2p-railing": "~0.7.1",
-    "libp2p-secio": "~0.8.1",
+    "libp2p-railing": "~0.7.0",
+    "libp2p-circuit": "0.0.2",
+    "libp2p-secio": "~0.8.0",
     "libp2p-spdy": "~0.11.0",
     "libp2p-tcp": "~0.11.0",
-    "libp2p-webrtc-star": "~0.13.2",
+    "libp2p-webrtc-star": "~0.13.0",
     "libp2p-websockets": "~0.10.1",
     "lodash.times": "^4.3.2",
     "pre-commit": "^1.2.2",
     "pull-goodbye": "0.0.2",
     "pull-serializer": "^0.3.2",
-    "pull-stream": "^3.6.1",
+    "pull-stream": "^3.6.0",
     "safe-buffer": "^5.1.1",
     "sinon": "^2.3.6",
     "electron-webrtc": "^0.3.0",
     "wrtc": "0.0.62"
   },
-  "dependencies": {
-    "async": "^2.5.0",
-    "libp2p-ping": "~0.4.0",
-    "libp2p-swarm": "~0.29.1",
-    "libp2p-circuit": "~0.0.1",
-    "mafmt": "^2.1.8",
-    "multiaddr": "^2.3.0",
-    "peer-book": "~0.4.0",
-    "peer-id": "~0.8.7",
-    "peer-info": "~0.9.2"
-  },
   "contributors": [
     "Chris Bratlien <chrisbratlien@gmail.com>",
-    "Daijiro Wachi <daijiro.wachi@gmail.com>",
     "David Dias <daviddias.p@gmail.com>",
     "Elven <mon.samuel@qq.com>",
     "Friedel Ziegelmayer <dignifiedquire@gmail.com>",
