+++ test/node.js
@@ -8,4 +8,4 @@ require('./nodejs-bundle/stream-muxing')
 require('./nodejs-bundle/discovery')
 require('./nodejs-bundle/peer-routing')
 require('./nodejs-bundle/content-routing')
-require('./nodejs-bundle/circuit/circuit')
+require('./nodejs-bundle/circuit')
