+++ test/nodejs-bundle/circuit/index.js
@@ -1,0 +1,5 @@
+'use strict'
+
+require('./dial-over-any-relay')
+require('./dial-over-specific-relay')
+require('./dial-over-specific-relay-transport')
